+++ ContentPatcher/Framework/Conditions/ConditionType.cs
@@ -207,6 +207,9 @@ namespace ContentPatcher.Framework.Conditions
         /// <summary>The <see cref="StardewValley.LocalizedContentManager.LanguageCode"/> name.</summary>
         Language,
 
+        /// <summary>The current content pack's unique manifest ID.</summary>
+        ModId,
+
         /****
         ** Specialized
         ****/